Reformatting the code so that the client and server code exist separately
