+++ 25 Foodies Starting Project/app/meals/share/page.jsx
@@ -1,23 +1,10 @@
 import classes from './page.module.css';
 import ImagePicker from '@/components/meals/image-picker';
 
+// Reformatting the code ensures that server side and client side code are separate
+import { shareMeal } from '@/lib/actions';
+
 export default function ShareMealPage() {
-	async function shareMeal(formData) {
-		// Need to add this directive to execute on the server
-		'use server';
-
-		const meal = {
-			title: formData.get('title'),
-			summary: formData.get('summary'),
-			instructions: formData.get('instructions'),
-			image: formData.get('image'),
-			creator: formData.get('name'),
-			creator_email: formData.get('email'),
-		};
-
-		console.log(meal);
-	}
-
 	return (
 		<>
 			<header className={classes.header}>
