+++ 25 Foodies Starting Project/lib/actions.jsx
@@ -1,0 +1,17 @@
+'use server';
+
+export async function shareMeal(formData) {
+	// Need to add this directive to execute on the server
+	'use server';
+
+	const meal = {
+		title: formData.get('title'),
+		summary: formData.get('summary'),
+		instructions: formData.get('instructions'),
+		image: formData.get('image'),
+		creator: formData.get('name'),
+		creator_email: formData.get('email'),
+	};
+
+	console.log(meal);
+}
